requirements.txt
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
packaging==24.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pypdfium2==4.30.0
pytesseract==0.3.13
